유저와 상품을 분리하기 위해 패키지 구조 변경
