--- app/src/main/java/com/codesoom/assignment/ProductNotFoundException.java
@@ -1,7 +1,0 @@
-package com.codesoom.assignment;
-
-public class ProductNotFoundException extends RuntimeException {
-    public ProductNotFoundException(Long id) {
-        super(id + "를 찾을 수 없습니다.");
-    }
-}
