--- app/src/main/java/com/codesoom/assignment/domain/ProductRepository.java
@@ -1,18 +1,0 @@
-package com.codesoom.assignment.domain;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface ProductRepository {
-    List<Product> findAll();
-
-    Optional<Product> findById(Long id);
-
-    Product save(Product product);
-
-    void delete(Product product);
-
-    void deleteAll();
-
-    void deleteById(Long id);
-}
