--- app/src/main/java/com/codesoom/assignment/domain/Status.java
@@ -1,5 +1,0 @@
-package com.codesoom.assignment.domain;
-
-public enum Status {
-    SALE, SOLD_OUT
-}
