--- app/src/main/java/com/codesoom/assignment/dto/ErrorResponse.java
@@ -1,13 +1,0 @@
-package com.codesoom.assignment.dto;
-
-public class ErrorResponse {
-    private String message;
-
-    public ErrorResponse(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
