--- app/src/main/java/com/codesoom/assignment/dto/ValidationFailResponse.java
@@ -1,14 +1,0 @@
-package com.codesoom.assignment.dto;
-
-import lombok.Getter;
-
-@Getter
-public class ValidationFailResponse {
-    private final String fieldName;
-    private final String message;
-
-    public ValidationFailResponse(String fieldName, String message) {
-        this.fieldName = fieldName;
-        this.message = message;
-    }
-}
