+++ app/src/main/java/com/codesoom/assignment/product/application/ToyCommandService.java
@@ -1,0 +1,42 @@
+package com.codesoom.assignment.product.application;
+
+import com.codesoom.assignment.product.dto.ProductNotFoundException;
+import com.codesoom.assignment.product.domain.Product;
+import com.codesoom.assignment.product.domain.ProductRepository;
+import com.codesoom.assignment.product.dto.ListToDelete;
+import com.codesoom.assignment.product.dto.ProductData;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+public class ToyCommandService implements ProductCommandService {
+    private final ProductRepository productRepository;
+
+    public ToyCommandService(ProductRepository productRepository) {
+        this.productRepository = productRepository;
+    }
+
+    @Override
+    public Product create(ProductData productData) {
+        return productRepository.save(productData.toProduct());
+    }
+
+    @Override
+    public Product update(Long id, ProductData data) {
+        return productRepository.findById(id)
+                .orElseThrow(() -> new ProductNotFoundException(id))
+                .change(data.toProduct());
+    }
+
+    @Override
+    public void deleteById(Long id) {
+        productRepository.deleteById(id);
+    }
+
+    @Override
+    public void deleteAllByList(ListToDelete list) {
+        list.getIdList()
+                .forEach(this::deleteById);
+    }
+}
