+++ app/src/main/java/com/codesoom/assignment/product/domain/Product.java
@@ -1,0 +1,91 @@
+package com.codesoom.assignment.product.domain;
+
+import lombok.Builder;
+import lombok.Getter;
+
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import java.util.Objects;
+
+@Entity
+@Getter
+@Builder
+public class Product {
+    @Id
+    @GeneratedValue
+    private Long id;
+
+    private String name;
+
+    private String maker;
+
+    private Integer price;
+
+    private String imageUrl;
+
+    @Enumerated(EnumType.STRING)
+    private Status status;
+
+    public Product() {}
+
+    public Product(String name, String maker, Integer price, String imageUrl) {
+        this.name = name;
+        this.maker = maker;
+        this.price = price;
+        this.imageUrl = imageUrl;
+    }
+
+    public Product(Long id, String name, String maker, Integer price, String imageUrl) {
+        this.id = id;
+        this.name = name;
+        this.maker = maker;
+        this.price = price;
+        this.imageUrl = imageUrl;
+    }
+
+    public Product(Long id, String name, String maker, Integer price, String imageUrl, Status status) {
+        this.id = id;
+        this.name = name;
+        this.maker = maker;
+        this.price = price;
+        this.imageUrl = imageUrl;
+        this.status = status;
+    }
+
+    public Product change(Product product) {
+        this.name = product.getName();
+        this.maker = product.getMaker();
+        this.price = product.getPrice();
+        this.imageUrl = product.getImageUrl();
+        this.status = product.getStatus();
+        return this;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+
+        Product product = (Product) o;
+
+        return Objects.equals(id, product.id) &&
+                Objects.equals(name, product.name) &&
+                Objects.equals(maker, product.maker) &&
+                Objects.equals(price, product.price) &&
+                Objects.equals(imageUrl, product.imageUrl) &&
+                status == product.status;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, name, maker, price, imageUrl, status);
+    }
+}
