+++ app/src/main/java/com/codesoom/assignment/product/domain/ProductRepository.java
@@ -1,0 +1,18 @@
+package com.codesoom.assignment.product.domain;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface ProductRepository {
+    List<Product> findAll();
+
+    Optional<Product> findById(Long id);
+
+    Product save(Product product);
+
+    void delete(Product product);
+
+    void deleteAll();
+
+    void deleteById(Long id);
+}
