+++ app/src/main/java/com/codesoom/assignment/product/domain/Status.java
@@ -1,0 +1,5 @@
+package com.codesoom.assignment.product.domain;
+
+public enum Status {
+    SALE, SOLD_OUT
+}
