+++ app/src/main/java/com/codesoom/assignment/product/dto/ProductNotFoundException.java
@@ -1,0 +1,7 @@
+package com.codesoom.assignment.product.dto;
+
+public class ProductNotFoundException extends RuntimeException {
+    public ProductNotFoundException(Long id) {
+        super(id + "를 찾을 수 없습니다.");
+    }
+}
