+++ app/src/main/java/com/codesoom/assignment/product/infra/JpaProductRepository.java
@@ -1,0 +1,23 @@
+package com.codesoom.assignment.product.infra;
+
+import com.codesoom.assignment.product.domain.Product;
+import com.codesoom.assignment.product.domain.ProductRepository;
+import org.springframework.context.annotation.Primary;
+import org.springframework.data.repository.CrudRepository;
+
+import java.util.List;
+import java.util.Optional;
+
+@Primary
+public interface JpaProductRepository
+        extends ProductRepository, CrudRepository<Product, Long> {
+    List<Product> findAll();
+
+    Optional<Product> findById(Long id);
+
+    Product save(Product product);
+
+    void delete(Product product);
+
+    void deleteAll();
+}
